+++ src/main/java/com/bship/games/domains/Game.java
@@ -1,4 +1,4 @@
-package com.bship.games.models;
+package com.bship.games.domains;
 
 import java.util.List;
