+++ src/main/java/com/bship/games/domains/Point.java
@@ -1,4 +1,4 @@
-package com.bship.games.models;
+package com.bship.games.domains;
 
 public class Point {
     private int x;
