+++ src/test/java/com/bship/games/repositories/BoardRepositoryTest.java
@@ -1,7 +1,7 @@
 package com.bship.games.repositories;
 
-import com.bship.games.models.Board;
-import com.bship.games.models.Game;
+import com.bship.games.domains.Board;
+import com.bship.games.domains.Game;
 import org.apache.tomcat.jdbc.pool.DataSource;
 import org.flywaydb.core.Flyway;
 import org.junit.Before;
